Move into its own file and rename to make clear this is not a cleanup task
